first commit
There's a Jupyter notebook that goes through the process of preparing
and writing three JSON files, in `final_data`.

Mathematical letters are mapped to their standard counterparts, with
annotations for the mathematical styling, e.g. BOLD or DOUBLE-STRUCK.

For everything in the category `Sm` (Mathematical symbols), I went
through and wrote out mappings to conventional operator symbols or
function names, along with some annotations.

Finally, for Greek letters, there is a mapping to their Latin names.